// src/modules/ng-date-input/src/interfaces/IDateRangePickerOptions.ts
import { ValidatorFn } from '@angular/forms';
import * as momentNs from 'moment';

import { IDefinedDateRange } from './IDefinedDateRange';

export interface IDateRangePickerOptions {
	autoApply?: boolean;
	clickOutsideAllowed?: boolean;
	displayFormat?: string;
	disabled?: boolean;
	disableInputDisplay?: boolean;
	format: string;
	icons?: 'default' | 'material' | 'font-awesome';
	minDate?: momentNs.Moment;
	maxDate?: momentNs.Moment;
	position?: 'left' | 'right';
	preDefinedRanges?: IDefinedDateRange[];
	showResetButton?: boolean;
}